# docs/v2/customer/customer_swagger.yaml
definitions:
...
  customer.EntitiesResponseTemplate-customer_debtResp:
    properties:
      count:
        type: integer
      next:
        type: string
      previous:
        type: string
      results:
        items:
          $ref: '#/definitions/customer.debtResp'
        type: array
    type: object
...
  customer.changePasswordReq:
    properties:
      confirm_password:
        type: string
      old_password:
        type: string
      password:
        type: string
    required:
    - confirm_password
    - old_password
    - password
    type: object
...
  customer.debtCancelReq:
    properties:
      description:
        type: string
    required:
    - description
    type: object
  customer.debtCreateReq:
    properties:
      amount:
        type: number
      description:
        type: string
      receiver_id:
        type: string
    required:
    - amount
    - description
    - receiver_id
    type: object
  customer.debtFulfillReq:
    properties:
      otp:
        type: string
      token:
        type: string
    required:
    - otp
    - token
    type: object
  customer.debtFulfillResp:
    properties:
      token:
        type: string
    type: object
  customer.debtResp:
    properties:
      amount:
        type: number
      create_time:
        type: string
      description:
        type: string
      id:
        type: string
      owner_bank_account_number:
        type: string
      owner_bank_name:
        type: string
      owner_id:
        type: string
      owner_name:
        type: string
      receiver_bank_account_number:
        type: string
      receiver_bank_name:
        type: string
      receiver_id:
        type: string
      receiver_name:
        type: string
      status:
        $ref: '#/definitions/debt.Status'
      transaction_id:
        type: string
      update_time:
        type: string
    type: object
  customer.errorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      message:
        type: string
    type: object
...
  customer.meResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  debt.Status:
    enum:
    - pending
    - pending
    - cancelled
    - fulfilled
    type: string
    x-enum-varnames:
    - DefaultStatus
    - StatusPending
    - StatusCancelled
    - StatusFulfilled
...
info:
  contact: {}
paths:
...
  /api/customer/v1/customers/{id}:
    get:
      consumes:
      - application/json
      description: Get a customer
      operationId: customer-get
      parameters:
      - description: ID of customer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.meResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a customer
      tags:
      - Customer
...
  /api/customer/v1/me/:
    get:
      consumes:
      - application/json
      description: Get profile
      operationId: me
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.meResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get profile
      tags:
      - Me
...
  /api/customer/v1/me/change-password:
    post:
      consumes:
      - application/json
      description: Change password
      operationId: change-password
      parameters:
      - description: Change password
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.changePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.meResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      summary: Change password
      tags:
      - Me
...
  /api/customer/v1/me/debts:
    get:
      consumes:
      - application/json
      description: Show debt
      operationId: debt-listing
      parameters:
      - description: ID of bank account
        in: query
        name: owner_id
        type: string
      - description: ID of bank account
        in: query
        name: receiver_id
        type: string
      - description: Status of debt
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.EntitiesResponseTemplate-customer_debtResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Show debt
      tags:
      - Debt
    post:
      consumes:
      - application/json
      description: Create a debt
      operationId: debt-create
      parameters:
      - description: Create a debt
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.debtCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/customer.debtResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Create a debt
      tags:
      - Debt
  /api/customer/v1/me/debts/{id}:
    get:
      consumes:
      - application/json
      description: Get a debt
      operationId: debt-get
      parameters:
      - description: ID of debt
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.debtResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Get a debt
      tags:
      - Debt
  /api/customer/v1/me/debts/cancel/{id}:
    put:
      consumes:
      - application/json
      description: Cancel a debt
      operationId: debt-cancel
      parameters:
      - description: ID of debt
        in: path
        name: id
        required: true
        type: string
      - description: Cancel a debt
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.debtCancelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.debtResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Cancel a debt
      tags:
      - Debt
  /api/customer/v1/me/debts/fulfill-with-token/{id}:
    put:
      consumes:
      - application/json
      description: Fulfill a debt with token
      operationId: fulfill-debt-with-token
      parameters:
      - description: ID of debt
        in: path
        name: id
        required: true
        type: string
      - description: Fulfill a debt with token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/customer.debtFulfillReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.debtResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      summary: Fulfill a debt with token
      tags:
      - Debt
  /api/customer/v1/me/debts/fulfill/{id}:
    put:
      consumes:
      - application/json
      description: Fulfill a debt
      operationId: debt-fulfill
      parameters:
      - description: ID of debt
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/customer.debtFulfillResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customer.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customer.errorResponse'
      security:
      - Bearer: []
      summary: Fulfill a debt
      tags:
      - Debt
...
swagger: "2.0"